+++ CHANGES.md
@@ -5,6 +5,7 @@
 ## New Features
 
 - Updated CMake files to use CMakeSDK v2.0 or greater
+- Add `magic_enum` to `VarAPI` to convert `enum class` to/from string and to iterate all values
 
 
 # Version 1.4.0